Modularize decorator classes
