+++ decorator.rb
@@ -1,0 +1,14 @@
+require './nameable'
+
+class Decorator < Nameable
+  attr_accessor :nameable_object
+
+  def initialize(nameable_object)
+    super()
+    @nameable_object = nameable_object
+  end
+
+  def correct_name
+    @nameable_object.correct_name
+  end
+end
